Modernize old test/lib/statemachine_test. It runs! It passes!
